Separate the eventer into interfaces
