+++ internal/controlplane/server_test.go
@@ -67,7 +67,7 @@ func init() {
 	// It would be nice if we could Close() the httpServer, but we leak it in the test instead
 }
 
-func newDefaultServer(t *testing.T, mockStore *mockdb.MockStore) *Server {
+func newDefaultServer(t *testing.T, mockStore *mockdb.MockStore) (*Server, events.Interface) {
 	t.Helper()
 
 	evt, err := events.Setup(context.Background(), &serverconfig.EventConfig{
@@ -89,7 +89,7 @@ func newDefaultServer(t *testing.T, mockStore *mockdb.MockStore) *Server {
 
 	server, err := NewServer(mockStore, evt, c, mockJwt)
 	require.NoError(t, err, "failed to create server")
-	return server
+	return server, evt
 }
 
 func generateTokenKey(t *testing.T) string {
